Moved FeaturedSlide to shared directory
