+++ frontend/foodieland-app/components/shared/FeaturedSlider.tsx
